48450: We should not remove "FieldDeclaration" of root component.

+++ org.eclipse.wb.core/src/org/eclipse/wb/internal/core/model/variable/FieldInitializerVariableSupport.java
@@ -264,6 +264,9 @@ public final class FieldInitializerVariableSupport extends FieldVariableSupport 
   ////////////////////////////////////////////////////////////////////////////
   @Override
   public void deleteAfter() throws Exception {
+    if (m_javaInfo.isRoot()) {
+      return;
+    }
     delete_removeDeclarationField();
   }
